Extract selected outline keyboard controls
